fix: after creating the contract there are always success/error CompleteModal
