+++ src/components/CompleteModal/CompleteModal.tsx
@@ -1,8 +1,9 @@
 import React, { useMemo, VFC } from 'react';
 
-import clsx from 'clsx';
-import { Modal } from 'components/Modal';
 import { Box, Typography } from '@material-ui/core';
+import clsx from 'clsx';
+
+import { Modal } from 'components/Modal';
 import { SuccessBigIcon, ErrorBigIcon } from 'theme/icons';
 import { useStyles } from './CompleteModal.styles';
