+++ src/pages/LostKeyContractPreview/LostKeyContractPreview.tsx
@@ -1,5 +1,5 @@
 import React, {
-  Fragment, useCallback, useEffect, useMemo, useState,
+  Fragment, useCallback,
 } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
@@ -8,21 +8,19 @@ import clsx from 'clsx';
 import Web3 from 'web3';
 
 import {
-  Preview, Copyable, FullscreenLoader, CompleteModal,
+  Preview, Copyable,
 } from 'components';
 import { useShallowSelector } from 'hooks';
 import {
-  ILostKeyContract, ILostKeyContractDynamicForm, RequestStatus, State, UserState,
+  ILostKeyContract, ILostKeyContractDynamicForm, State, UserState,
 } from 'types';
 import { routes } from 'appConstants';
 import contractFormsSelector from 'store/contractForms/selectors';
-import uiSelector from 'store/ui/selectors';
 import user from 'store/user/selectors';
 import { deleteLostKeyContractForm } from 'store/contractForms/reducer';
 import { getDeepValueByPath } from 'utils';
 import { useWalletConnectorContext } from 'services';
 import { createLostKeyContract } from 'store/contractForms/actions';
-import actionTypes from 'store/contractForms/actionTypes';
 import {
   staticLostKeyContractPreviewHelpers,
 } from './LostKeyContractPreview.helpers';
@@ -55,46 +53,6 @@ export const LostKeyContractPreview = () => {
   const lostKeyContract = useShallowSelector<State, ILostKeyContract>(
     contractFormsSelector.getLostKeyContract,
   );
-  const createContractRequestStatus = useShallowSelector(
-    uiSelector.getProp(actionTypes.CREATE_LOSTKEY_CONTRACT),
-  );
-  const isLoader = useMemo(
-    () => createContractRequestStatus === RequestStatus.REQUEST,
-    [createContractRequestStatus],
-  );
-
-  const [resultModalState, setResultModalState] = useState({
-    open: false,
-    result: false,
-  });
-  const handleCloseResultModal = useCallback(() => {
-    setResultModalState({
-      ...resultModalState,
-      open: false,
-    });
-  }, [resultModalState]);
-
-  useEffect(() => {
-    switch (createContractRequestStatus) {
-      case RequestStatus.SUCCESS: {
-        setResultModalState({
-          open: true,
-          result: true,
-        });
-        break;
-      }
-      case RequestStatus.ERROR: {
-        setResultModalState({
-          open: true,
-          result: false,
-        });
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-  }, [createContractRequestStatus]);
 
   const classes = useStyles();
 
@@ -198,12 +156,6 @@ export const LostKeyContractPreview = () => {
           ))}
         </Box>
       ))}
-      {isLoader && <FullscreenLoader />}
-      <CompleteModal
-        open={resultModalState.open}
-        result={resultModalState.result}
-        onClose={handleCloseResultModal}
-      />
     </Preview>
   );
 };
