+++ src/pages/TokenContractPreview/TokenContractPreview.tsx
@@ -1,9 +1,6 @@
 import React, {
   Fragment,
   useCallback,
-  useEffect,
-  useMemo,
-  useState,
 } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
@@ -15,22 +12,18 @@ import {
   Preview,
   YesNoBlock,
   Copyable,
-  FullscreenLoader,
-  CompleteModal,
 } from 'components';
 import { useShallowSelector } from 'hooks';
 import contractFormsSelector from 'store/contractForms/selectors';
-import uiSelector from 'store/ui/selectors';
 import user from 'store/user/selectors';
 import {
-  RequestStatus, State, TokenContract, UserState,
+  State, TokenContract, UserState,
 } from 'types';
 import { routes } from 'appConstants';
 
 import { deleteTokenContractForm } from 'store/contractForms/reducer';
 import { useWalletConnectorContext } from 'services';
 import { createTokenContract } from 'store/contractForms/actions';
-import actionTypes from 'store/contractForms/actionTypes';
 import { useStyles } from './TokenContractPreview.styles';
 import {
   dynamicTokenContractPreviewHelpers,
@@ -64,46 +57,6 @@ export const TokenContractPreview = () => {
   const tokenContract = useShallowSelector<State, TokenContract>(
     contractFormsSelector.getTokenContract,
   );
-  const createContractRequestStatus = useShallowSelector(
-    uiSelector.getProp(actionTypes.CREATE_TOKEN_CONTRACT),
-  );
-  const isLoader = useMemo(
-    () => createContractRequestStatus === RequestStatus.REQUEST,
-    [createContractRequestStatus],
-  );
-
-  const [resultModalState, setResultModalState] = useState({
-    open: false,
-    result: false,
-  });
-  const handleCloseResultModal = useCallback(() => {
-    setResultModalState({
-      ...resultModalState,
-      open: false,
-    });
-  }, [resultModalState]);
-
-  useEffect(() => {
-    switch (createContractRequestStatus) {
-      case RequestStatus.SUCCESS: {
-        setResultModalState({
-          open: true,
-          result: true,
-        });
-        break;
-      }
-      case RequestStatus.ERROR: {
-        setResultModalState({
-          open: true,
-          result: false,
-        });
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-  }, [createContractRequestStatus]);
 
   const classes = useStyles();
   let totalTokenAmount = 0;
@@ -232,12 +185,6 @@ export const TokenContractPreview = () => {
           </Fragment>
         );
       })}
-      {isLoader && <FullscreenLoader />}
-      <CompleteModal
-        open={resultModalState.open}
-        result={resultModalState.result}
-        onClose={handleCloseResultModal}
-      />
     </Preview>
   );
 };
